please see description - wonky as hell
So, for some reason the database does not appear to work correctly. I
believe it has previously worked properly, so this is new. Trying to get
gender with !gender will enter through all the callbacks correctly.
Setting your gender with !gender will not work correctly, as it skips a
callback. After this point, you can not get any data from the database.
If you restart, those changes get committed to disk and then you can
access them at next reboot, but this is not correct.